reconfigure all braintree payment code to no longer use ValidateMethod
