+++ imports/plugins/included/braintree/server/methods/braintreeapi.js
@@ -1,24 +1,30 @@
 /* eslint camelcase: 0 */
 // meteor modules
 import { Meteor } from "meteor/meteor";
-import { SimpleSchema } from "meteor/aldeed:simple-schema";
 // reaction modules
 import { Packages } from "/lib/collections";
 import { Reaction, Logger } from "/server/api";
 import Future from "fibers/future";
 import Braintree from "braintree";
+import accounting from "accounting-js";
 
-export const BraintreeApi = {};
-BraintreeApi.methods = {};
 
-export const refundDetailsSchema = new SimpleSchema({
-  transactionId: { type: String },
-  amount: { type: Number, decimal: true }
-});
+function getPaymentObj() {
+  return {
+    amount: "",
+    options: {submitForSettlement: true}
+  };
+}
 
-export const refundListSchema = new SimpleSchema({
-  transactionId: { type: String }
-});
+function parseCardData(data) {
+  return {
+    cardholderName: data.name,
+    number: data.number,
+    expirationMonth: data.expirationMonth,
+    expirationYear: data.expirationYear,
+    cvv: data.cvv
+  };
+}
 
 
 function getSettings(settings, ref, valueName) {
@@ -76,54 +82,53 @@ getRefundDetails = function (refundId) {
 };
 
 
-// BraintreeApi.methods.createCharge = new ValidatedMethod({
-//   name: "BraintreeApi.methods.createCharge",
-//   validate: new SimpleSchema({
-//     refundDetails: { type: refundDetailsSchema } //UPDATE
-//   }).validator(),
-//   run({ refundDetails }) { //UPDATE
-//   }
-// });
+export function paymentSubmit(paymentSubmitDetails){
+  let gateway = getGateway();
+  let paymentObj = getPaymentObj();
+  if (paymentSubmitDetails.transactionType === "authorize") {
+    paymentObj.options.submitForSettlement = false;
+  }
+  paymentObj.creditCard = parseCardData(paymentSubmitDetails.cardData);
+  paymentObj.amount = paymentSubmitDetails.paymentData.total;
+  let fut = new Future();
+  gateway.transaction.sale(paymentObj, Meteor.bindEnvironment(function (error, result) {
+    if (error) {
+      fut.return({
+        saved: false,
+        error: error
+      });
+    } else if (!result.success) {
+      fut.return({
+        saved: false,
+        response: result
+      });
+    } else {
+      fut.return({
+        saved: true,
+        response: result
+      });
+    }
+  }, function (error) {
+    Reaction.Events.warn(error);
+  }));
+
+  return fut.wait();
+}
 
 
-// BraintreeApi.methods.captureCharge = new ValidatedMethod({
-//   name: "BraintreeApi.methods.captureCharge",
-//   validate: new SimpleSchema({
-//     refundDetails: { type: refundDetailsSchema } //UPDATE
-//   }).validator(),
-//   run({ refundDetails }) { //UPDATE
-//   }
-// });
+export function captureCharge(paymentCaptureDetails) {
+  let transactionId = paymentCaptureDetails.transactionId;
+  let amount = accounting.toFixed(paymentCaptureDetails.amount, 2);
+  let gateway = getGateway();
+  const fut = new Future();
 
-
-BraintreeApi.methods.createRefund = new ValidatedMethod({
-  name: "BraintreeApi.methods.createRefund",
-  validate: new SimpleSchema({
-    refundDetails: { type: refundDetailsSchema }
-  }).validator(),
-  run({ refundDetails }) {
-    let transactionId = refundDetails.transactionId;
-    let amount = refundDetails.amount;
-    let gateway = getGateway();
-    const fut = new Future();
-    gateway.transaction.refund(transactionId, amount, Meteor.bindEnvironment(function (error, result) {
+  if (amount === accounting.toFixed(0, 2)) {
+    gateway.transaction.void(transactionId, function (error, result) {
       if (error) {
         fut.return({
           saved: false,
           error: error
         });
-      } else if (!result.success) {
-        if (result.errors.errorCollections.transaction.validationErrors.base[0].code === "91506") {
-          fut.return({
-            saved: false,
-            error: "Braintree does not allow refunds until transactions are settled. This can take up to 24 hours. Please try again later."
-          });
-        } else {
-          fut.return({
-            saved: false,
-            error: result.message
-          });
-        }
       } else {
         fut.return({
           saved: true,
@@ -131,38 +136,99 @@ BraintreeApi.methods.createRefund = new ValidatedMethod({
         });
       }
     }, function (e) {
-      Logger.fatal(e);
-    }));
+      Logger.warn(e);
+    });
     return fut.wait();
   }
-});
+  gateway.transaction.submitForSettlement(transactionId, amount, Meteor.bindEnvironment(function (error, result) {
+    if (error) {
+      fut.return({
+        saved: false,
+        error: error
+      });
+    } else {
+      fut.return({
+        saved: true,
+        response: result
+      });
+      //This is here for ease while testing, and will be removed before it's live
+      //This is here for ease while testing, and will be removed before it's live
+      //This is here for ease while testing, and will be removed before it's live
+      //This is here for ease while testing, and will be removed before it's live
+      gateway.testing.settle(transactionId, function (err, settleResult) {
+        settleResult.success
+        // true
+
+        settleResult.transaction.status
+        // Transaction.Status.Settled
+      });
+      //This is here for ease while testing, and will be removed before it's live
+      //This is here for ease while testing, and will be removed before it's live
+      //This is here for ease while testing, and will be removed before it's live
+      //This is here for ease while testing, and will be removed before it's live
+    }
+  }, function (e) {
+    Logger.warn(e);
+  }));
+
+  return fut.wait();
+}
 
 
-BraintreeApi.methods.listRefunds = new ValidatedMethod({
-  name: "BraintreeApi.methods.listRefunds",
-  validate: new SimpleSchema({
-    refundListDetails: { type: refundListSchema }
-  }).validator(),
-  run({ refundListDetails }) {
-    let transactionId = refundListDetails.transactionId;
-    let gateway = getGateway();
-    this.unblock();
-    let braintreeFind = Meteor.wrapAsync(gateway.transaction.find, gateway.transaction);
-    let findResults = braintreeFind(transactionId);
-    let result = [];
-    if (findResults.refundIds.length > 0) {
-      for (let refund of findResults.refundIds) {
-        let refundDetails = getRefundDetails(refund);
-        result.push({
-          type: "refund",
-          amount: parseFloat(refundDetails.amount),
-          created: moment(refundDetails.createdAt).unix() * 1000,
-          currency: refundDetails.currencyIsoCode,
-          raw: refundDetails
+export function createRefund(refundDetails) {
+  let transactionId = refundDetails.transactionId;
+  let amount = refundDetails.amount;
+  let gateway = getGateway();
+  const fut = new Future();
+  gateway.transaction.refund(transactionId, amount, Meteor.bindEnvironment(function (error, result) {
+    if (error) {
+      fut.return({
+        saved: false,
+        error: error
+      });
+    } else if (!result.success) {
+      if (result.errors.errorCollections.transaction.validationErrors.base[0].code === "91506") {
+        fut.return({
+          saved: false,
+          error: "Braintree does not allow refunds until transactions are settled. This can take up to 24 hours. Please try again later."
+        });
+      } else {
+        fut.return({
+          saved: false,
+          error: result.message
         });
       }
+    } else {
+      fut.return({
+        saved: true,
+        response: result
+      });
     }
+  }, function (e) {
+    Logger.fatal(e);
+  }));
+  return fut.wait();
+}
 
-    return result;
+
+export function listRefunds(refundListDetails) {
+  let transactionId = refundListDetails.transactionId;
+  let gateway = getGateway();
+  let braintreeFind = Meteor.wrapAsync(gateway.transaction.find, gateway.transaction);
+  let findResults = braintreeFind(transactionId);
+  let result = [];
+  if (findResults.refundIds.length > 0) {
+    for (let refund of findResults.refundIds) {
+      let refundDetails = getRefundDetails(refund);
+      result.push({
+        type: "refund",
+        amount: parseFloat(refundDetails.amount),
+        created: moment(refundDetails.createdAt).unix() * 1000,
+        currency: refundDetails.currencyIsoCode,
+        raw: refundDetails
+      });
+    }
   }
-});
+
+  return result;
+}
